requirements.txt
Pillow~=8.2.0
uuid~=1.30